Use JedisPool to provide Jedis instances
Apparently Jedis is not designed to share across threads. Thus we use
the recommended option: JedisPool. This thus removes the ugly critical
region hack.
This should also solve the issue of Redis not being ready when system
or container restart. From my understanding to the Jedis impl, Jedis
only attempts to connect to Redis server when JedisPool.getResource()
is called, effectively being lazy-evaluated.